Revert int cast removals

+++ src/header/time/mod.rs
@@ -256,7 +256,8 @@ pub extern "C" fn clock() -> clock_t {
     }
     let ts = unsafe { ts.assume_init() };
 
-    let clocks = ts.tv_sec * CLOCKS_PER_SEC + (ts.tv_nsec / (1_000_000_000 / CLOCKS_PER_SEC));
+    #[expect(clippy::unnecessary_cast, reason="needed on i586")]
+    let clocks = ts.tv_sec * CLOCKS_PER_SEC as i64 + (ts.tv_nsec / (1_000_000_000 / CLOCKS_PER_SEC)) as i64;
     clock_t::try_from(clocks).unwrap_or(-1)
 }
